requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
contentful==1.6.0
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-dateutil==2.7.2
python-dotenv==0.8.2
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
